Refactor segmentation model preparation by extracting to function Set default AdInpaintWidth and AdInpaintHeight for Face/Person model. Also use set Sampler and CfgScale
This commit extracts the logic for preparing segmentation model parameters to a separate function, SegmModelParameters. Previously, this logic was located in queue.go as part of a block of code constructing an ADeatiler instance. This chunk of code was becoming increasingly complex and was a good candidate for refactoring. Moving this logic to a separate function makes the code clearer and easier to maintain. Changes to how segmentation models are handled can now be made in one place and it's easier to test and verify these changes.