Ensure AlwaysOnScripts is initialized before use
In the file 'imagine_queue/queue.go' there could be a panic due to a nil pointer dereference when trying to access the ADetailer field of a possibly non-initialized AlwaysOnScripts field. To prevent this, additional code has been added to initialize AlwaysOnScripts, if it is nil, before trying to access its ADetailer field.