Changed the behavior of processImagineSDModelNameSetting() handler
Now we don't use the method UpdateModelName that was deprecated in dd1c973f7f2646ada7373156f20b67de2bf08c55, and we grab the current model when settingsMessageComponents() is called which appends it to the checkpoint selection list (674c4c00c8cfd2fafe85d669c7bac1cfcaf76bda with fixes from 04b3182d5b612aaa17cba85ec9b269fb0bf856ab)

UpdateModelName is deprecated as that only changed the model when the next generation queue is ran, as previously each generation would read from the DefaultSettings to then update the model.

We also update the message once a checkpoint is selected to inform that a model change is currently happening, then temporarily change it to "Updated model to %v", which changes back to the default after 5 seconds.