Fix: tooltip parsing with regex
Now we use FindAllStringSubmatch and use the last [][]string from the matches to check if we have a matching weight.

This fixes a long issue when a user types in a lora that matches the pattern in the different OR `|` pipes. e.g. `concept\\loraName.safetensors:0.35` would fail, because it was expecting loraName:0.35 instead.

This fix addresses all the combinations of folder name being present, .safetensors or .ckpt, etc. and now parses the weight correctly