Continue deprecating ImageGeneration and TextToImageRequest structs
Continuation from e1135cd52c85c291d236ae28d97e1e8981b065a7, we implement and move over to the new structs.

ImageGeneration from image_generation.go becomes ImageGenerationRequest, which inherits TextToImageRequest from text_to_image.go. This is so that in case we need to add new fields, we only need to edit in one place. The TextToImageRequest is the struct that creates the final JSON to be passed to the txt2img api