Initial implementation of controlnet options
Included options to autocomplete the preprocessor and model list and attachment

Moved DenoisingStrength to Img2ImgItem and added a new ControlnetItem to the queue. This is so that we can separate the image attachment from each

Changed the String(int) string and Len() int in the Cacheable interface to inherit from fuzzy.Source instead

Added a controlnetTypes() method that's unused for now to dynamically update the available types. To be deprecated as it invalidates the slash command on restart as it recreates the choices without consistency at the moment

Modified the ControlnetParameters struct to accept ControlMode and ResizeMode aliased types for future use

Removed the `input =` and the return for the autocomplete helper methods, as we're not using them elsewhere.

Calculate extraLoras by using `min(extraLoras, 25-len(options))`, as well as sanity checks for the len(options), as we can only have a maximum of 25

We've reduced the amount of extraLoras to 2, (total of 3 including the default)

The controlnet.go file in stable_diffusion_api does multiple things. To be able to easily access the various types, modules (preprocessors) and models, we populate those list when we all the main ControlnetTypes.apiGet()

We maintain a separate list on the top level struct as Modules and Models. They contain the available Models and Modules to them respectively

We grab the key and store it as the type for both of them