Fix: Img2Img as controlnet
We've removed embedding the thumbnail while we're still generating, as it's challenging to modify the attached files afterward.

We've also removed adding the img2img to the controlnet script if controlnet is enabled, but a controlnet image is not included. Controlnet will automatically use img2img as its input

Moved appending of primary image in imageEmbedFromAttachment to the if `image != nil` block, so that we can only put in a thumbnail if desired.

Renamed thumbnails to thumbnailBuffers in generate.go and image_to_image.go

Attempt to remove empty thumbnailBuffers in generate.go and image_to_image.go

Adjusted the time.Since to only parse if it's non zero, and rounded to seconds.

Only print ADetailer scripts as Controlnet potentially have a very long base64 byte buffer

Added a new tiler called compositor that can take an arbitrary aspect ratio to attempt to tile