Implement: Interrupt button
We've added a new progress response that has all of the fields from the api, such as the current image, job state, etc

Using a channel, we receive a new Interaction from c.Interrupt and store it in c.DiscordInteraction

RemoveFromQueue now renamed the interaction to a messageInteraction to clearly state that we're looking for an interaction's message's interaction (thanks discord)

For now we do the checking by using the mentions slice. Because we always store who ran a command, re-roll, or variation button, we can use this to compare who pressed the button (i.Member.User.ID)

We've also added components with their disabled counterparts in components.go and generate.go

We store the message returned to us by EditInteractionResponse and store it in queue.DiscordInteraction.Message, if it is nil, in command-handlers.go

Added checks for generationEmbedDetails in embed_helpers.go to append "Interrupted" with a boolean

Also fixed delete button by replacing the CustomID with the one in handlers.DeleteGeneration to always ensure using the correct string, aslong with upscale, re-roll and variations.

At the moment, deleting a re-roll doesn't work, as Discord doesn't give complete information about the interaction. Future implementation might work if we store the interaction ID or the valid ID for deletion in the delete button itself, and handle it that way.