Fix regression: Batch count not being stored
Because of Batchount in image_generation.go from 58a315f5de1d3ab016a43e8f129088b7ccb1cd07 getting added, sqlite.go had the wrong batch count field getting set and recorded

Because of this, previous fields had batch count set to 0, so we've added a check for minimum NIter to 1 in generate.go and image_to_image.go

Fixed imageEmbedFromBuffers in embed_helpers.go not appending the only embed when there's no primary image. Now we immediately append the first embed immediately and mutate it for the rest if there are images or thumbnails present. This allows us to have the embed even if we don't have images.

Also we've modified the empty message as the final embed to just include a <@ID> mention to help with mention checking when attempting to delete the message

We've also pre-added the embed in the webhook using `&[]*discordgo.MessageEmbed{embed}` to ensure the webhook is present, despite imageEmbedFromBuffers mutating the webhook to add this later on