Implement cfg rescale scripts as a manual --parameter
We've moved individual script methods and arg bodies to their respective files

QueueItem gets an entities.Scripts, future rewrite can just include this object to construct what we need

Also added the ability to add a {DEBUG} at the end of the prompt to print to the console

interfaceConvertAuto now returns the ValueType interface conversion regardless if field *F is nil or not. It just doesn't attempt to write to *F if nil.

Append scripts in embed with a []string, and also display number of total images

Removed checking of AlwaysonScripts individual nilness