Complete implementation of /raw command
We've added a new method called TextToImageRaw so that we can pass in a raw []byte request.

We unmarshal to the default queue item first, then attach the pointer to queue.Raw. Then we unmarshal again to the raw's TextToImageRequest.

Potentially we could explore unmarshalling only to raw's TextToImageRequest, but then we wouldn't be able to write to the inner scripts for other parts of the code base.

Along with this, we've added a new field in QueueItem as Raw of *entities.TextToImageRaw, which inherits *TextToImageRequest and a RawScripts of map[string]any. Crucially, this RawScripts overrides the scripts in TextToImageRequest to potentially any scripts even if we have not implemented them yet.

We've also added a check for BatchSize to be a minimum of 1 and max of 4 after processing.