Add raw json file upload to /raw command
The /raw command now can accept a json file uploaded directly. If no json file is uploaded, useDefault gets put into modalDefault variable as a map[string]bool

We unmarshal to the default queue item first, then attach the pointer to queue.Raw. Then we unmarshal again to the raw's TextToImageRequest.

Fixed bug of ItemTypeRaw to now shadow err variable in text_to_image.go

Moved jsonToQueue as a separate method that accepts the interaction, useDefault boolean and the jsonBlob as a []byte

Unmarshalling to TextToImageRequest might fail so we still continue in this case.