Fix regression: Getting params fcb6bfc9a74fba65bbb1069a1d244be50421f43a
If err != nil, we shouldn't be reading from params.

Additionally, fixed ImageGenerationRequest being nil if we're not using a default queue item.

Also fixed checking if TextToImageRequest is nil before putting it in queue.Raw by immediately creating the ImageGenerationRequest.

Because Unmarshal most likely assigns and writes to a new address, we need to write back to queue.ImageGenerationRequest.TextToImageRequest

We've also put in the defaultWidth and defaultHeight for the DefaultQueueItem() method

Future implementation should record the output response after inference to use the values provided by the stable diffusion API in cases where parameters are left blank. (e.g., Sampler is an empty string in the database, but the API will return the sampler used regardless)